separate shortcode grammar and parsing utils
